ai-lang-gragh/pyproject.toml
[project]
name = "ai-lang-gragh"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-openai>=0.3.25",
    "langgraph>=0.4.9",
    "python-dotenv>=1.1.1",
]
